Add support for Extension Refs and make Refs easier to reason about.
An interface T Extends another interface U if T's method set is a
superset of U's method set. This can be verified by testing that T is
assignable to U.

Refs are now easier to reason about. A ref has a RefType, a source and
a destination. A source or a destination means a Position (a
goref.Position, not a token.Position), a Package and an Identifier.