Treat package imports as just another type of Ref.
Since refs have a position, this means creating a lot more Refs: each
file importing another package will have a ref to every file in the
imported package. The FromPosition is the import statement, the
FromIdentifier is either the name qualifying the import or the name of
the imported package ("." is just like any other name here). The
ToPosition is the position in each file of the package clause.