Rename PackageGraph.LoadPrograms to LoadPackages.
Long ago, goref assumed that the entrypoint packages would be 'main'
packages, i.e. that whole programs would be indexed at a time. It
turned out to be an arbitrary restriction with no benefits, so any
package can now be loaded. As a result, the function needed to be
renamed.